refactor(storage): extract write/delete operations into storage/mutation.rs
Moves all DB mutation functions from mod.rs (370 → 120 lines) into a new
mutation.rs (269 lines). Tightens visibility: insert_chunk_row →
pub(crate), insert_sub_embeddings → pub(super), delete_file_chunks_in →
pub(crate). Loop inserts switch to prepare_cached; DELETEs use CTE to
share the subquery across all six statements. Public API preserved via
re-exports in mod.rs.

Closes #45
